token-exchange/pyproject.toml
[project]
name = "identity-broker"
version = "0.1.0"
description = "Vault Identity Broker — exchanges a Vault token for a Vault-signed OIDC Identity JWT"
requires-python = ">=3.12"
dependencies = [
    "cachetools>=7.0.3",
    "fastapi>=0.135.1",
    "hvac>=2.4.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "requests>=2.32.0",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "api",
    "app_logging",
    "broker",
    "config",
    "exceptions",
    "models",
    "verify",
]
